+++ src/pages/report/params.js
@@ -3,10 +3,6 @@ import * as base64 from '../../core/utils/base64.js';
 const knownEncodeParams = new Set(['query', 'view', 'pipeline', 'title', 'dzen', 'noedit']);
 const knownDecodeParams = new Set(['q', 'v', 'p', 'title', 'dzen', 'noedit']);
 
-function encodeSearchParamPair(name, value) {
-    return encodeURIComponent(name) + '=' + encodeURIComponent(value);
-}
-
 function ensureString(value, fallback) {
     return typeof value === 'string' ? value : fallback || '';
 }
@@ -22,18 +18,18 @@ export function encodeParams(params) {
         extra
     } = typeof params === 'string' ? { query: params } : params;
     const pipeline = [];
-    const parts = [];
+    const pairs = [];
 
     if (dzen) {
-        parts.push('dzen');
+        pairs.push(['dzen']);
     }
 
     if (noedit) {
-        parts.push('noedit');
+        pairs.push(['noedit']);
     }
 
     if (title) {
-        parts.push(encodeSearchParamPair('title', title));
+        pairs.push(['title', title]);
     }
 
     // backward compatibility
@@ -54,22 +50,23 @@ export function encodeParams(params) {
         const pipelineJson = JSON.stringify(pipeline);
 
         if (pipelineJson !== '[["query",""],["view",null]]') {
-            parts.push(encodeSearchParamPair('p', base64.encode(pipelineJson)));
+            pairs.push(['p', base64.encode(pipelineJson)]);
         }
     }
 
     if (extra) {
         Object.keys(extra).sort().forEach(name => {
             if (!knownEncodeParams.has(name)) {
-                parts.push(encodeSearchParamPair(name, extra[name]));
+                pairs.push([name, extra[name]]);
             }
         });
     }
 
-    return parts.join('&');
+    return pairs;
 }
 
-export function decodeParams(params) {
+export function decodeParams(pairs) {
+    const params = Object.fromEntries(pairs);
     const query = base64.decode(ensureString(params.q, '')); // backward compatibility
     const view = 'v' in params ? base64.decode(ensureString(params.v, '')) : undefined; // backward compatibility
     const rawPipeline = JSON.parse(base64.decode(ensureString(params.p, '')) || '[]');