+++ src/views/editors.js
@@ -47,17 +47,15 @@ class Editor extends Emitter {
             mode: 'javascript',
             theme: 'neo',
             indentUnit: 0,
-            hintOptions: autocomplete && {
+            showHintOptions: {
                 hint: autocomplete,
-                completeSingle: false,
-                closeOnUnfocus: true,
                 isolateStyleMarker: this.getIsolateStyleMarker()
             }
         });
 
         cm.on('change', () => this.emit('change', cm.getValue()));
 
-        if (autocomplete) {
+        if (typeof autocomplete === 'function') {
             // patch prepareSelection to inject a context hint
             // const ps = cm.display.input.prepareSelection;
             // cm.display.input.prepareSelection = function(...args) {
@@ -68,8 +66,8 @@ class Editor extends Emitter {
             //     return selection;
             // };
 
-            cm.on('cursorActivity', cm => cm.state.focused && cm.showHint(autocomplete));
-            cm.on('focus', cm => !cm.state.completionActive && cm.showHint(autocomplete));
+            cm.on('cursorActivity', cm => cm.state.focused && cm.showHint());
+            cm.on('focus', cm => !cm.state.completionActive && cm.showHint());
         }
 
         this.cm = cm;