+++ src/views/input.js
@@ -1,7 +1,7 @@
 /* eslint-env browser */
 
-import defined from '../core/utils/defined.js';
 import safeFilterRx from '../core/utils/safe-filter-rx.js';
+import debounceFn from '../core/utils/debounce.js';
 
 export default function(discovery) {
     const factories = {
@@ -10,10 +10,25 @@ export default function(discovery) {
     };
 
     discovery.view.define('input', function(el, config, data, context) {
-        const { name, value, type = 'text', placeholder, onInit, onChange, htmlType = 'text', htmlMin, htmlMax } = config;
+        const {
+            name,
+            value,
+            type = 'text',
+            placeholder,
+            onInit,
+            onChange,
+            htmlType = 'text',
+            htmlMin,
+            htmlMax,
+            debounce
+        } = config;
         const factory = factories[type] || factories.text;
         const inputEl = el.appendChild(document.createElement('input'));
-        let lastInput = defined([value, context[name]], '');
+        let lastInput = value ? discovery.query(value, data, context) : context[name];
+
+        if (typeof lastInput !== 'string') {
+            lastInput = '';
+        }
 
         inputEl.type = htmlType;
         inputEl.value = lastInput;
@@ -30,20 +45,20 @@ export default function(discovery) {
             inputEl.max = htmlMax;
         }
 
-        inputEl.addEventListener('input', () => {
+        inputEl.addEventListener('input', debounceFn(() => {
             const newInput = inputEl.value.trim();
 
             if (lastInput !== newInput) {
                 lastInput = newInput;
 
                 if (typeof onChange === 'function') {
-                    onChange(factory(newInput), name);
+                    onChange(factory(newInput), name, data, context);
                 }
             }
-        });
+        }, debounce));
 
         if (typeof onInit === 'function') {
-            onInit(factory(inputEl.value.trim()), name);
+            onInit(factory(inputEl.value.trim()), name, data, context);
         }
     });
 }